Started working on the ball class. Have created ball, called ball.render and called ball.move in Game. But I am having issues with getting ball.move to work. Work in progress

+++ src/com/oose/breakout/Entity.java
@@ -8,12 +8,12 @@ public class Entity {
 	protected int positionY;
 	protected Image entityImage;
 	protected boolean isAlive;
-	public int speed;
+	public float speed;
 	
 	
 	
 	//Setters & Getters of entity
-	public int getSpeed() {
+	public float getSpeed() {
 		return speed;
 	}
 	public void setSPEED(int _speed) {
